+++ common/src/main/java/dev/terminalmc/chatnotify/ChatNotify.java
@@ -18,10 +18,7 @@ package dev.terminalmc.chatnotify;
 
 import com.mojang.datafixers.util.Pair;
 import dev.terminalmc.chatnotify.compat.commandkeys.CommandKeysWrapper;
-import dev.terminalmc.chatnotify.config.Config;
-import dev.terminalmc.chatnotify.config.Notification;
-import dev.terminalmc.chatnotify.config.ResponseMessage;
-import dev.terminalmc.chatnotify.config.Trigger;
+import dev.terminalmc.chatnotify.config.*;
 import dev.terminalmc.chatnotify.util.ModLogger;
 import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
@@ -66,6 +63,7 @@ public class ChatNotify {
         for (Notification notif : config.getNotifs()) {
             for (Trigger trig : notif.triggers) {
                 if (trig.type == Trigger.Type.REGEX) trig.tryCompilePattern();
+                if (trig.styleTarget.type == StyleTarget.Type.REGEX) trig.styleTarget.tryCompilePattern();
             }
             for (Trigger trig : notif.exclusionTriggers) {
                 if (trig.type == Trigger.Type.REGEX) trig.tryCompilePattern();