+++ common/src/main/java/dev/terminalmc/chatnotify/gui/widget/list/option/NotifOptionList.java
@@ -67,9 +67,9 @@ public class NotifOptionList extends OptionList {
             } else {
                 addEntry(new Entry.TriggerFieldEntry(dynEntryX, dynEntryWidth, entryHeight,
                         this, notif, trigger, i));
-                if (trigger.styleString != null) {
-                    addEntry(new Entry.StyleStringFieldEntry(dynEntryX, dynEntryWidth, entryHeight,
-                            this, trigger));
+                if (trigger.styleTarget.enabled) {
+                    addEntry(new Entry.StyleTargetFieldEntry(dynEntryX, dynEntryWidth, entryHeight, 
+                            this, trigger.styleTarget));
                 }
             }
         }
@@ -170,7 +170,7 @@ public class NotifOptionList extends OptionList {
         // Check whether the drop location is valid
         if (hoveredEntry instanceof Entry.TriggerFieldEntry || hoveredSlot == triggerListOffset() - 1) {
             // pass
-        } else if (hoveredEntry instanceof Entry.StyleStringFieldEntry) {
+        } else if (hoveredEntry instanceof Entry.StyleTargetFieldEntry) {
             hoveredSlot -= 1; // Reference the 'parent' Entry
         } else {
             this.dragSourceSlot = -1;
@@ -252,7 +252,7 @@ public class NotifOptionList extends OptionList {
                                 (button) -> {
                                     this.setDragging(true);
                                     list.dragSourceSlot = list.children().indexOf(this);
-                                    list.dragHasStyleField = trigger.styleString != null;
+                                    list.dragHasStyleField = trigger.styleTarget.enabled;
                                 })
                         .pos(x - list.smallWidgetWidth - SPACING, 0)
                         .size(list.smallWidgetWidth, height)
@@ -316,15 +316,15 @@ public class NotifOptionList extends OptionList {
                 // Style string add button
                 Button styleButton = Button.builder(Component.literal("+"),
                                 (button) -> {
-                                    trigger.styleString = "";
+                                    trigger.styleTarget.enabled = true;
                                     list.reload();
                                 })
                         .pos(movingX, 0)
                         .size(list.tinyWidgetWidth, height)
                         .build();
-                if (trigger.styleString == null) {
+                if (!trigger.styleTarget.enabled) {
                     styleButton.setTooltip(Tooltip.create(
-                            localized("option", "notif.restyle_string.add.tooltip")));
+                            localized("option", "notif.style_target.add.tooltip")));
                     styleButton.setTooltipDelay(Duration.ofMillis(500));
                 } else {
                     styleButton.active = false;
@@ -344,12 +344,11 @@ public class NotifOptionList extends OptionList {
             }
         }
 
-        private static class StyleStringFieldEntry extends Entry {
-            StyleStringFieldEntry(int x, int width, int height, NotifOptionList list,
-                                  Trigger trigger) {
+        private static class StyleTargetFieldEntry extends Entry {
+            StyleTargetFieldEntry(int x, int width, int height, NotifOptionList list,
+                                  StyleTarget styleTarget) {
                 super();
-                int fieldSpacing = 1;
-                int stringFieldWidth = width - (list.tinyWidgetWidth * 3) - (fieldSpacing * 2);
+                int stringFieldWidth = width - (list.tinyWidgetWidth * 4);
                 int movingX = x + list.tinyWidgetWidth;
 
                 // Info icon
@@ -357,18 +356,36 @@ public class NotifOptionList extends OptionList {
                         Component.literal("\u2139"), Minecraft.getInstance().font);
                 infoIcon.alignCenter();
                 infoIcon.setTooltip(Tooltip.create(
-                        localized("option", "notif.restyle_string.tooltip")));
+                        localized("option", "notif.style_target.tooltip")));
                 infoIcon.setTooltipDelay(Duration.ofMillis(500));
                 elements.add(infoIcon);
-                movingX += list.tinyWidgetWidth + fieldSpacing;
+                movingX += list.tinyWidgetWidth;
+
+                // Type button
+                CycleButton<StyleTarget.Type> typeButton = CycleButton.<StyleTarget.Type>builder(
+                                (type) -> Component.literal(type.icon))
+                        .withValues(StyleTarget.Type.values())
+                        .displayOnlyValue()
+                        .withInitialValue(styleTarget.type)
+                        .withTooltip((type) -> Tooltip.create(
+                                localized("option", "notif.style_target.type." + type + ".tooltip")))
+                        .create(movingX, 0, list.tinyWidgetWidth, height, Component.empty(),
+                                (button, type) -> {
+                                    styleTarget.type = type;
+                                    list.reload();
+                                });
+                typeButton.setTooltipDelay(Duration.ofMillis(500));
+                elements.add(typeButton);
+                movingX += list.tinyWidgetWidth;
 
                 // Style string field
                 TextField stringField = new TextField(movingX, 0, stringFieldWidth, height);
+                if (styleTarget.type == StyleTarget.Type.REGEX) stringField.regexValidator();
                 stringField.setMaxLength(240);
-                stringField.setValue(trigger.styleString);
-                stringField.setResponder((string) -> trigger.styleString = string.strip());
+                stringField.setValue(styleTarget.string);
+                stringField.setResponder((string) -> styleTarget.string = string.strip());
                 stringField.setTooltip(Tooltip.create(
-                        localized("option", "notif.restyle_string.field.tooltip")));
+                        localized("option", "notif.style_target.field.tooltip")));
                 stringField.setTooltipDelay(Duration.ofMillis(500));
                 elements.add(stringField);
                 movingX = x + width - list.tinyWidgetWidth;
@@ -376,7 +393,7 @@ public class NotifOptionList extends OptionList {
                 // Delete button
                 elements.add(Button.builder(Component.literal("\u274C"),
                                 (button) -> {
-                                    trigger.styleString = null;
+                                    styleTarget.enabled = false;
                                     list.reload();
                                 })
                         .pos(movingX, 0)