+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed a bug causing exclusion triggers to be checked even when disabled
 - Fixed a crash caused by processing invalid message structures
 - Fixed trigger editor restyle preview not working for key-type triggers
+- Added option to use regex for style target
 
 ## 2.3.2